chore: cleanup run code
